requirements.txt
requests~=2.32.3
pytest~=8.3.5
allure-python-commons~=2.14.0
attrs == 25.3.0
certifi == 2025.1.31
charset-normalizer == 3.4.1
colorama == 0.4.6
idna == 3.10
iniconfig == 2.1.0
packaging == 24.2
pip == 25.0.1
pluggy == 1.5.0
pytest == 8.3.5
requests == 2.32.3
urllib3 == 2.3.0
